Fix: Plugin Next Page Going to 404
It happened after adding priority option in the `Route` class. Because
now the `manager/plugin` is on priority 10, `manager/plugin/(:num)` is
automatically on priority 10.1 and `manager/plugin/(:any)` —that should
be on priority 10.2— is on priority 10.

At least I have to set the last route in a larger priority because this
→ `if(File::exist(PLUGIN . DS . '{{page offset}}' . DS . 'launch.php')`
… will never exist, that’s why we are always directed to 404 page. It
could be easily solved by adding a larger priority to the last route but
it may also messed up another routes. For now I fixed the problem with
this regex pattern: `([^0-9\/]+)` which means “not a page steps”.